Handle page content/url
remove
