+++ front/pages/api/v1/w/[wId]/assistant/conversations/index.ts
@@ -24,6 +24,18 @@ import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { apiError } from "@app/logger/withlogging";
 
+// TODO Move this to request body
+const jitActions = [
+  {
+    id: -1,
+    sId: "jit",
+    type: "request_user_data_configuration" as const,
+    name: "brower_context",
+    description: "This will return contextual information about the browser",
+    available_data: ["page_content", "page_url"],
+  },
+];
+
 /**
  * @swagger
  * /api/v1/w/{wId}/assistant/conversations:
@@ -221,6 +233,7 @@ async function handler(
               profilePictureUrl: message.context.profilePictureUrl,
               origin: message.context.origin ?? "api",
             },
+            jitActions,
           },
           { resolveAfterFullGeneration: blocking === true }
         );
